Add peer discovery
A method for letting peers advertise each other's existence is added to
the peer_loop. This is used spread knowledge about peers throughout the
network.

I wasn't yet able to test this functionality though, as the `peer_loop`
method is not yet written in a generic way using traits. The
`serialized` abstract argument should have its type replaced with a
generic type argument conforming to some specific traits, so we can
build tests of this method.