Verify network on peer connection
Add a check that the peers run the same network before connecting. This
should avoid a very common user error where a user accidently connects
to a peer on the wrong network. This check was missing from some
Bitcoin Lightning nodes until recently, and that was quite annoying for
me.

This closes #12.