Remove `extra_values` from handshake data
We don't need this for the node upgrades to be backwards compatible
as we can just add new verbs to the peer communication instead. This
means that the exchange of new data introduced in an upgrade will
not be exchanged through a handshake but through peer requests and
responses instead.

This keeps the handshake logic and data structure simpler.