Add special case for when block is found by running client
If a client finds a block by mining, this block should be shared with
all peers immediately without going through the pull process where a
block is instead advertised by informing the peer that a block with a
certain height and hash has been found. In other words: when block is
found by running peer: send block to all peers; when block is shared
from other peer: inform peers about new block height and block hash and
let them request it, if its new for them.