Bind the result of `broadcast::channel`
This fixes #17.

The cause of #17 is not actually that the miner channel was prematurely closed,
but that the peer broadcast channel was.

As soon as all receivers of a broadcast channel have been dropped, the channel
itself is dropped and all subsequent sends and receives fail. Because we were
not binding the initial receiver, the channel was immediately closed after being
created.

The reason why we didn't catch this was that we handled the error like a
user-error, returning a `Result`, where instead this should be a fatal error, as
it is caused by our incompetence and not the users. I've amended the error
handling to make sure that it doesn't happen again in the future.