Initial sketch of RPC implementation using tarpc
This probably shouldn't be merged as is, but it can serve as a discussion point
going forward. In fact, I'm not even sure I think tarpc is a good choice for
this.

To try it out, run the server in one terminal:

```
cargo run
```

Then connect to it from another terminal

```
cargo run --bin rpc -- --server-addr 127.0.0.1:23455 BlockHeight
```

Currently, the rate-limiter is bugged such that subsequent rpc calls will
sometimes fail. I'm not sure what's wrong.