Add database initialization to startup process
Two `levelDB` are created upon program initialization where each
database is equivalent to a table in SQL and is a key/value map
stored on disk. One database maps block height to block hash, whereas
the other database maps block hash to blocks.

These databases may have mutual consistency requirements where a
change made to both databases must be atomic. For this reason, the
database object is wrapped in a mutex and a atomic reference counter
(Arc) to ensure atomic updates.

The main thread will definitely need to read and write to the database
but will the peer threads also need this? Maybe they just need to read
from the database, or maybe they don't need to read or write from it at
all?