Send atomic database objects to peer threads
With this commit peer threads receive not only a hash map of peers but
also an object for atomic database access such that they can read (and
write) atomically from the database. We probably only want the peers to
read from the database and only have the main thread write to the
database. Perhaps this can be implemented through traits later down the
line.

Speaking of traits ... I tried to implement the `State` object that is
passed around with a trait for the `Database` access instead of using
the `Database` type directly, but I ran into problems of the sizes not
being known at compile time, lifetime requirements and more, which I was
unable to solve at the time. So I implemented `State` with the
`Database` type instead. This meant that I had to create on-disk
databases when the unit tests are running, and I had to perform a neat
little trick to make them run in parallel.

On Linux the unit test databases are created in the `/tmp` folder, so it
should be OK to clutter there a bit, but we probably want to clean up
after ourselves regardless? This hasn't been implemented in the tests
yet! There is probably some syntax for running functions after the
execution of all tests.