Replace tracking of latest state with database
Instead of tracking the latest block height with an in-memory object, we
track the latest state with a new database (a levelDB key/value map).
While we are still developing the skeleton, we wish to keep the state
management as simple as possible, and I believe this is the best way to
achieve this.

This means that the state is automatically preserved across a shutdown
of a client, since the `latest` database only has *one* entry that is
read whenever needed. It also means that a new block will result in
three write operations to the databases instead of two as previously.

A simple optimization that can be made in the future would be to add an
in-memory element to state such that the latest block height can be
read without locking all databases as we are currently doing.

This also changes the Mutex type on `peer_map` from tokio to std since
the tokio version is only needed when we are awaiting across a lock, and
we are not doing that for the `peer_map` data structure.