Restructure file hierarchy for models
Instead of having one big file with a flat structure, we attempt to put
the files into logically named files. The `config_models` directory
cannot be merged with the `models` directory though, since the former is
used both in `lib.rs` and in `main.rs` whereas the latter is only used
in `lib.rs`.