Move main loop to separate file
This commit moves the "main loop" which was previously contained in the
`lib.rs/initialize` function into a separate file. It also renames the
`mine` file into `mine_loop` as it is technically a loop that can either
find a block or get a new block from `main_loop`. This way, we have
separate files for `peer_loop`, `main_loop`, and `mine_loop`. In the
future, we will add an `rpc_loop` for handling RPC calls.