Verify received block is new in main loop
Repeat the check that the block is new in `main_loop`, as a block could
have been received and inserted between a peer thread's check and the
handling in `main_loop`.

Before this fix, the same block was saved to the database multiple
times as the peer threads all verified that the block was new, and only
afterwards did the main thread handle all the insertions. This problem
was revealed by the associated `run-multiple-instances-advanced.sh`
script.