Ensure peers do not connect to self or to already connected peers
We introduce a `peer_id` in the handshake which is unique for a running
instance of the program, and is generated at startup. This ID ensures
that a client program does not connect to itself and that it does not
connect to the same peer twice (as could happen since an ingoing and
outgoing connection to the same peer would otherwise not be spotted).

This closes #30.