Simplify MS' `prove`
Removes a conditional statement that is only relevant if one of the bit
indices are less than window start for the item about to be added to the
mutator set.

I think the reason that this piece of code can be removed is that the
bit indices for this membership proof will always live in the active
window when it is inserted. So it chunk dictionary will be empty. The
chunk dictionary is only populated as elements are added to the mutator
set *after* the current one has been inserted.