Refactor MS' `verify` to use a hash map for bit indices
We don't have a benchmark for mutator sets, but this speeds up the tests
by 50 %: On my Lenovo P1 all mutator set tests took 28s before this change
and the take 18s after this change, even though this change adds a new
test. Note, however, that this speedup is not observed when the
`NUM_TRIALS` is very low. The above time was for `NUM_TRIALS = 160`
which is what we are using in production.

Note that with this commit, the constant running time for `verify` is
ditched. I don't think a constant running time is needed. We might as
well short circuit, as the input or the state are not secret.