Add archival mutator set
This archival mutator set uses a archival MMRs for the AOCL and SWBF
commitments. It also stores the chunks seperately since the MMRs only
contain digests.

Next up we should define a trait that both the "regular" mutator set and
the archival mutator set can implement. This way, both can be used with
the same function names, but the archival one will have more methods.