Improve structure of `update_from_remove`, add return value
The return value indicates if the membership proof has changed or not.
It's honestly an optimization that adds some complexity. So it might not
be a good idea, but it should reduce the amount of I/O in applications
using this library.

I wrote this before I learned to do pointer book keeping in Rust. With
this commit, a lot of copying is eliminated.