Change chunk and active window to use byte arrays
Replace their use of bool arrays with byte arrays. This is done for two
reasons: Secondarily because it's better style and takes up less RAM.
Primarily, because the deserialization of some data structures (set
commitment) was failing prior to this change.

Cf.
https://stackoverflow.com/questions/72618777/how-to-deserialize-a-nested-big-array

A solution to this Stack Overflow question is simply to change the data
type.