Separate digest type into three digest types
I tried to make the `Digest` just a type alias for the array
`[BFieldElement; 6]` but then I had some conversion problems, as I can't
implement a conversion from e.g. `Digest` to `[u8; 48]` which I need for
e.g. network and database serialization. Therefore I instead defined a
struct as `pub struct Digest([BFieldElement; 6])`. And then I added a
digest type that can be used as a key in a database and one that defines
an order which can be used for the PoW algorithm.

These three digest types should possibly be just *one* digest type that
implements everything it needs to...

Cf. #4.