Change datatype for sending block from peer thread to main thread
With this commit peer threads communicate new blocks to the main thread by
sending a list of blocks, not by sending one block at a time. The reason
for this change is that sending one block at a time could lead to a race
condition where two peer threads have a different view of what's the
newest block and then sending these to the main thread, and the main
thread handling them out-of-order. This race condition could happen if
one of the peer threads send multiple blocks to the main thread and
these were not handled atomically but interrupted by a new block from
another thread which could have a higher PoW workload than the first in
the list of new blocks, but a lower than the last in the list. This
would result in the latter in the list not being stored, as they would
not necessarily have more PoW work than the next block in the list of
blocks being handled.