Reduce "state-by-instruction-pointer" from peer loop
This commit reduces state in the form of nested expectations of
call/response peer communication.

Instead of having code like this:
```rust
match msg {
  BlockNotification(x) => {
    peer.send(BlockRequestByHeight) => {
      match msg {
        BlockResponseByHeight(y) => {
          ...
        }
      }
    }
  }
}
```

we have code like this:
```rust
match msg {
  BlockNotification(x) => {
    peer.send(BlockRequestByHeight);
  },
  Block(y) => {
    ..
  }
}
```

So the response will be sendt as a `Block` message instead and will be
caught on the outer-most layer of the peer loop.

The reason for this change is that we have to accept other messages from
a peer while we are for example resolving a fork which includes many
messages back and forth to the peer.

Instead of storing the state of a peer-communication sequence as an
instruction pointer, this commit expands the state of the peer loop in
the form of the `PeerState` data structure. The field
`fork_reconciliation_blocks` is added to this data structure to handle a
fork reconciliation.