Allow fork reconciliation logic to be interrupted
This commit allows a fork reconciliation process to be interrupted with
other messages without breaking.

Also: if a block notification is received during a block reconciliation
event, then this block is requested once the block reconciliation is
completed. This is achieved by storing state in the peer state field
`highest_shared_block_height`.