Organize mutex/semaphors for databases into right categories
So far all databasees have shared the same mutex. With the introduction
of a database for IPs of banned peers, this is no longer appropriate
since we don't want to put a lock on the block databases when we check
if a peer should be or is banned.

It's worth noting though, that we don't need a lock per database (think
table, if you're used to SQL/relational databases) since some of them
are linked and should be opdated atomically. This concern is especially
relevant for block databases. Therefore we want to group the databases
so the relevant database sets share a mutex.