Refactor tests, move out of shared file
In rust tests usually live in the same file as the functions that are
tested. With this commit, that architecture is copied.

The tests in this project often involve a lot of setup/boilerplate
though, and this code now lives in a separate file: src/tests/shared.rs.

The RPC module should also be tested btw. A test module for rpc.rs
should be added.