Start work on synchronization mode
When the client is behind a peer by more than N blocks, the blocks
should be fetched sequentially, from lower to higher. A Special flag in
the state is used to indicate if the client is currently in this mode.

The current plan is that this synchronization will be handled inside of
`peer_loop`, in a tokio::select! macro such that other communications
with this peer are not blocked.

I'm satisfied if the missing blocks are fetched in-order. We don't need
to separate fetching and validation in order to be able to fetch faster.
Instead, let's just fetch everything in order and validate it before
requesting the next blocks. We can always improve on this logic
eventually.