Add peer_loop_wrapper for peer state management
Add a function called `peer_loop_wrapper` that calls the `peer_loop`
function but also handles the setup of state and the cleanup of state
after a peer is disconnected. This means that the logic that is
associated with a peer connection (reading their standing from the
database etc.) and that is shared both for incoming and outgoing
connections can be shared in one function and does not have to be
repeated as was the case prior to this commit.