Move genesis block to heap to avoid tokio-worker stack overflow
The archival state stores the genesis block to avoid having to
initialize it in each database lookup. But initializing it on the stack
lead to stack overflows in the tokio worker threads as documented in #32.
Therefore we move it from the stack to the heap by wrapping the object
in `Box`.

I was unable to capture the stack overflow problem in a test. But this
commit still contains a test that constructs an archival state inside of
a tokio-worker thread to give *some* assurance that this struct can be
constructed in this environment.