Change bincary serialization library for database values
Changing the serialization library to serde_stacker prevents a stack
overflow in the tokio-worker threads that are spun up to handle the
peer-loop for each peer.

Cf. #32.

We probably want to transition away from storing the blocks in the
database ASAP but for now I think we should merge the synchronization
logic into master and this change allows us to do that.