Change mutator set dependency to fix stack overflow problem
In the mutator set library, the active window is a [u32; 1000] data
structure. This does apparently not fit on the stack so it sometimes
crashes the entire program when being deserialized on the stack. The
implemented solution is to place that array on the heap instead. This
seems to have fixed the issue, so we can go back to the regular
serialization method that. See the parent of this commit.

This closes #32.