Avoid instance ID being set twice
Previously was two different instance IDs set for the same run of a
program. This is now avoided by setting it on the networking state and
always deriving the handshake data from the networking state.

This commits also removes the `async` from the method on `State` to
derive the handshake data.