ArchivalMutatorSet: Make remove() invertible
- Add `Option<Vec<u128>>` to remove() in MutatorSet trait
  that returns the flipped bits on ArchivalMutatorSet
- Calculate and return diff_indices in remove_helper()
- ArchivalMutatorSet: Add get_bloom_filter_bit()

In order to roll back a RemovalRecord from a Bloom filter, it is
necessary to record the bits that the RemovalRecord flips. This is only
possible for an ArchivalMutatorSet, since it contains the full MMR that
represents the Bloom filter.

A MutatorSetAccumulator must use another rollback strategy.

Planned: ArchivalMutatorSet: Add set_bloom_filter_bit()

Co-authored-by: Simon Shine <simon@neptune.cash>