Complete functionality to roll back mutator set
This rollback is needed when the archival node experiences a fork and
one or more blocks must be rolled back. Since an update to a Bloom
filter is not reversible, an extra piece of information must be stored
for each block: the indices of the bits that were flipped by this block.
We could probably have saved space by storing something else than just
the flipped bits (like the bits that the removal records of this block
would flip that were already flipped) but we thought that storing the
bits flipped by the block was the simplest approach.