Clarify that 2nd element of output is randomness, devnet
There was some confusion about the purpose of the 2nd element of the
output tuple. With this commit, we clarify that this element is the
randomness that unlinks the addition record from the removal record --
once we add the ZK-proof that the bit indices have been derived
correctly, that is.