Add batch update from addition for removal records
This allows us to update the removal records just like membership proofs
can be updated.

This functionality will be used when a block contains both addition and
removal records (output and input UTXOs, respectively). To apply these,
we need to update the removal records as the addition records are added.