Improve some comments in the archival block validation function
Most of the checks happen in the `devnet_is_valid`. `archival_is_valid`
only does the PoW check, then it calls the `devnet_is_valid` function.
This division is made such that we can also check the validity of block
templates, without having to generate a valid PoW threshold digest.