Add missing `reverse` to MS update's `apply` method
In the previous commit, I added the removal and the addition records in
the opposite order from how there were calculated in the tests and from
how they were calculated earlier. Now the addition/removal records are
applied from the beginning of the list such that element 0 is applied
first, then element 1 etc.