Fix connection race condition with extra check on peer map
Add a some checks on peer map prior to entering `peer_loop.run`: Verify
that instance ID for new peer is not found in peer map already and that
the peer's socket address is also not found in the peer map already.
Also check that we don't exceed max peer count.

These checks are performed earlier in the connection process but they
could become invalidated by other threads if e.g. an ingoing and an
outgoing connection from/to another peer is made simulataneously. This
race condition has been observed (by me) when spinning up 8 nodes on the
same computer.

This change had a rather big effect on the tests in peer_loop, so they
had to be rewritten. Turns out that they made some wrong assumptions
prior to this commit.