Fix another error in the update of the archival mutator set
Again the culprit was a reverse/missing reverse. In this commit an
aggressive assert is added. I bet we'll have fun with that: The assert
verifies that the block which is used to update the archival mutator set
has the same state of the mutator set as that in the block.