Switch around fields in `GlobalState` to fix lock acquisition ordering
Our policy is to acquire mutexes in the order that they are defined in
the struts in `GlobalState`. This way we should eliminate risks of
deadlocks. By switching around `Wallet` and `BlockchainState` we live up
to this politcy. The problem was that `BlockchainState` contains both a
std mutex and a tokio mutex and the latter requires an `await` to be
acquired, and std mutexes cannot be held over awaits.