Expand on tests of `get_connection_status`
Since this function receives `GlobalState` and checks if the incoming
connection is from a banned peer, it would be a shame to not test this
behavior as well.

We did already have a test that banned peers cannot make incoming
connections in `disallow_ingoing_connections_from_banned_peers_test` but
that test ran on a higher-level function. So this commit adds more
fine-grained testing of the important helper-function
`get_connection_status`.