Add handling of membership proofs to wallet
This function is called whenever a new block is received. It maintains
the membership proofs of all relevant (read: own) UTXOs for the last N
blocks where N is currently set to 500.

Note that membership proofs are not currently removed from the list of
maintained membership proofs once a UTXO is spent. That is a todo.
Another todo is to write tests of this complicated function
`update_wallet_state_with_new_block`.