Fix problem in wallet caused by multiple UTXOs having the same digest
If multiple monitored UTXOs in the wallet have the same digest and one
of these have to be marked as spent, we need to find the right one to
mark as spent. This is currently achieved on DevNet by using the
membership proof included in the block. On MainNet and TestNet we'll
have to come up with another solution for this funky cornercase. We'll
see once we get closer to the next iterations.