Change mutex type for light state to tokio
This is done to allow me to write the function to build a new
transaction. This function will require a lock on both the light state
and on the wallet state to ensure that no updates of the client's block
tip happens while the transaction is being built. If this were to
happen, a valid transaction could be returned.

Since Mutex from `std::sync` cannot be held over an `await` and
acquiring many of the locks for other fields in the `GlobalState`
struct requires an `await`, I felt forced to make this change.