Update mempool with block
This adds a function that updates the mempool with the content of a
block. The method first removes the transactions that have been included
or partly included in the latest block and then it updates the mutator
set data of the remaining transactions of the mempool.

The removal is done by comparing the removal records' bit indices of the
block and of the transactions in the mempool. If a transaction in the
mempool has a removal that matches a removal record from the block, then
this transaction is removed from the mempool as it would forever be
invalid (assuming no block reorganizations).

Thanks to Einar Rasmussen for coming up with a great design for the
mempool that I was able to use for this solution.