Rename 'Hashable' trait in neptune-core to 'Hashable2'
This trait conflicts in name with the trait in twenty-first. Since the
one in twenty-first plays together with RescuePrimeRegular via traits,
the Hashable2 trait is deemed to be replaced.

Since its .neptune_hash() function appears to be useful, this should be
considered in the Hasher trait, rather than create new traits with the
same names that compete for functionality.