Change interface to not return flipped Bloom filter values on remove
Previously the bit indices that were set by applying a removal record
was returned since these did not have to be the same as those from the
removal record (since a bit could only be set once). With the
introduction of the reversible Bloom filter in
97b9b518d66812daa1b12c846ae2f21b624aeef4 this return value is no longer
needed as the indices are appended to an index list and the difference
that the application of a removal record introduces is the same list of
indices as those contained in the removal record.

TL;DR: Reversible Bloom filters contain repeated indices, so we don't
need to track which bits a removal record flips.

Co-authored-by: Simon Shine <simon@neptune.cash>