Add check that mined block builds upon last canonical block
If a block is found quickly after one has been received by a peer, the
locally found block might not build upon the current blockchain tip. If
that's the case, we discard the locally found block and do *not* store
it to the database.

This check shouldn't affect miner efficiency negatively as the locally
mined block is always broadcast to all peers before being stored
locally.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>