Send request for block notification if peer signals heavier block
Upon connection to a new peer: If the peer signals a more canonical
(heavier in terms of PoW) block, then request a block notification
immediately, so we can catch up ASAP. This should speed up
synchronization if we're only a few blocks behind. The alternative would
be to wait for the peer to send a new block, and this new procedure is
simply much more efficient for catching up, when you're less than
`max-number-of-blocks-before-synching` behind the newly connected peer.