fix mining loop
The mining loop was comparing the block hash against the current
difficulty, whereas validity is defined relative to the previous
block's difficulty. This commit changes the mining loop to use
the difficulty that is provided as an argument, and where the
function is called the difficulty is supplied.

Also, the evolution of the nonce was entirely deterministic. This
commit makes it random (as per `StdRng`) in each iteration.

Also, the nonce was doubling as a counter to avoid checking a
termination condition in every iteration and instead checking it
only every 100 iterations. This function is now factored out into
a separate `counter` variable.