Remove `listen_address` from peer handshake
A peer does not know its own public IP address, so this should not be
part of the handshake. It only knows the port number that it is
accepting incoming connections on, so only this should be shared in the
handshake. A connection node, though, can derive which socket address a
peer is accepting incoming connections on, as it knows both the port and
the IP on which the connections was made.

This closes #9.