Add a test for correct handling of `get_latest_balance_height`
We can shoehorn a test of this function into an existing test by adding
a few asserts and verifying ensuring that also the light state is
updated correctly in the test.

This test can be used for test-driven development of (what I believe is)
a correct implementation of the method `get_latest_balance_height`.
Notice, though, that this test assumes that this method is a method on the
`GlobalState` struct and not, as previously, on wallet state.

Cf. #56.