refactor: rename sync locking methods
Renames locking methods in crate::sync to include the word "lock".

The word "with" is a bit vague, and when reading code that is using
the locks, it is helpful to clearly see where locks occur, and which
type.

The renames are:
 with()      --> lock()
 with_mut()  --> lock_mut()
 guard()     --> lock_guard()
 guard_mut() --> lock_guard_mut()

These match the naming in twenty_first::sync::*