Add logger output to slow startup check
A specific check that's performed on startup can take very long on slow
machines that have received a lot of UTXOs. This commit adds logger-output
such that the user can see how far along this process has come.

The startup check in question is the one that restores the monitored UTXO
data base from transaction recovery data stored on disk.