Align neptune-cli with RPC server
Ensure that names of commands match with those of the endpoints. Also
match ordering. It is still such that the RPC server has endpoints that
the CLI does nto include. That's OK since some of the RPC server
endpoints are only intended for the dashboard.

Also adds a CLI endpoint for returning the node's instance-ID.

Cf. #80.