feat: lock event callbacks
Atomic lock types in util_types::sync::tokio now accept
an optional name and lock-event callback fn.

This enables the app to be notified of lock events and take appropriate
action, such as logging/tracing.

These features are optional and the API is backwards compatible
with code that does not need lock-event notifications.