fix: prune_abandoned_monitored_utxos() RPC
WalletState::prune_abandoned_monitored_utxos() was causing 100% CPU
utilization due to calling futures::executor::block_on() on a tokio
async function, which is a no-no.

This patch uses `DbtVec::get_all()` and `::set_many()` instead of
`::iter_mut()`, which eliminates the core problem of holding a std::sync
lock over an `await`.

It also refactors the update section of the function, to be a bit more
readable/idiomatic.