fix: reduce dashboard CPU usage via fewer draws
The dashboard was constantly using around 15% CPU on my system while
doing basically nothing.  This drops usage to under 1%.

The high usage was caused by constantly re-drawing the screen in the
main loop.

The implemented fix is to add a `RefreshScreen` event which is sent
by sub-screens when data or key events occur that require a redraw.

In this way, redraws only occur when required.