refactor(block): Drop body hash from header
The block header does not need to commit to the block body. It is a
potential source of discrepancy and at best surplus weight.

Note that the block header does not commit to the body. However, the
proof-of-work puzzle is defined in terms of the *block's hash*,
which is `H(H(header)||H(body))`, so the *pair* is committed to when
the miner expends energy.