build: cargo update; fix tarpc/tokio_serde deps
After cargo update, we got tarpc build errors.

The cause was that tokio_serde had updated to a new major version and
neptune-core was using that version to interact with tarpc, which uses
older version.

The fix is to use tarpc::tokio_serde re-export.

The fix then exposed that we were not using the tarpc feature-flag
`serde-transport-json`.

With that corrected, it builds cleanly.