refactor: move cli args out of RwLock
The CLI args are immutable and do not need to be behind the
Global State lock.

This makes a read-only copy of the Cli Args in GlobalStateLock that
can be accessed without locking.

The read only property is enforced by making the field private and
adding a ::cli(&self) accessor method.  This is also done for the
GlobalLock struct.

Also adds a GlobalStateLock::set_cli() method that is only available
for test cases to simulate different CLI params.