fix: disable lock_event tracing to fix tests leak
The tracing-tests crate provides the #[traced_test] macro.

When used, log data gets written to a global static buffer.  See:
https://docs.rs/tracing-test/latest/tracing_test/

This was causing a huge mem leak, eg 6.5Gb+.

We may wish to re-evaluate use of #[traced_test], but for now
disabling lock_event tracing cures most of the leak.

Anyway, this tracing is very verbose and not usually needed.
we can easily re-enable if we need to debug a deadlock.