test: make test case less timing sensisitive
Test block_with_far_future_timestamp_is_invalid() would sometimes fail
when running tests in parallel because it depended on a sub-second
interval between when a value is set by the test-case and when it is
checked by Block::is_valid().

This commit makes the window 10 seconds, which hopefully should always
be enough.