refactor: Wrap witness UTXOs with salt
Witness objects such as `transaction::PrimitiveWitness` can contain
lists of input and output UTXOs. This commit wraps these lists of
into a new dedicated struct `SaltedUtxos` which adds a `salt` field.
The purpose is such that consensus programs can reduce a claim about
the set of input and/or output UTXOs to a claim about a single hash
digest. The salt ensures that claims related to different transactions
or blocks but happen to pertain to the same list of UTXOs (e.g., the
empty list) reduce to claims about distinct digests.