refactor: Expand trait `SecretWitness`
Trait `SecretWitness` is the interface between a blob of data that
makes an atomic unit of consensus logic valid, and the VM computation
induced by that validity claim. As such it supports the generation of
the program, the nondeterminism, and the input. The last one in that
list is the new function declarations.

Also in this commit: new trait `TypeScriptWitness` specializes
`SecretWitness` because for all type scripts the input and output are
the same. Specifically, the output is empty, and the input is the
triple (transaction kernel mast hash, salted input utxos hash, salted
output utxos hash).