style: rename crate::locks::sync to std
Under crate::locks we previously had:
 - sync
 - tokio

Each contains an impl of AtomicRw and AtomicMutex, with basically the
same API.  Yet:

 - sync refers to synchronous locks, ie sync vs async.
 - tokio refers to the tokio::sync lock implementation.

So the names are referring to different things.  Instead we change it to:

 - std
 - tokio

Now each refers to a lock implementation, ie std::sync and tokio::sync.

note: we could instead have changed `tokio` to `async`, but then there
might be multiple async lock impls to choose from.  So it seems cleanest
to use the name of each impl.