style: remove '_async' from ArchivalMmr methods
Previously ArchivalMmr async methods had the '_async' suffix to
differentiate from the Mmr trait's sync methods.

With the trait gone, we can revert back to the original method names
which are consistent with usage in MmrAccumulator.

Also, we do not use _async suffix for async methods anywhere else in
the codebase at this time.  (though maybe a good idea?)