fix: send syncing start/stop messages to miner
The mining loop is no longer async as it is wrapped by
spawn_blocking().

Thus it can no longer await async futures.  So we cannot check
the GlobalState syncing flag, which requires an async lock
acquisition.

Instead, the main task sends StartSyncing and StopSyncing
messages to the mining task which cancels the mining thread
when it receives StartSyncing.