perf: add db/storage bench tests
These are the bench tests from twenty_first::storage.

The db_dbtvec bench needed to be refactored to use async now that
storage_schema is entirely async.

The db_leveldb bench needed only changes to use statements.

It was also necessary to remove cfg(test) from an impl of
NeptuneLevelDb so that test methods are available for creating
a temporary database that will be cleaned up when dropped.