perf: speedup AMMR append operation
When appending a leaf to an MMR, we have the new leaf's authentication
path as we are building the MMR. So there's no reason to throw the
auth path away and then get the auth path later in a separate call.

This commit builds the auth path exactly once, and not twice as it was done
before. It also avoids unnecessary recursion.

This commit provides a more than 2x speedup for the mean and median case,
a 1.3x speedup for the slowest case and a more than 30x(!) speedup for cases
where the appends are followed by a `persist` to the DB.