refactor: Don't return peaks when getting MMR auth path from AMMR
No real noticable in performance, but this interface was not well liked,
so we remove the peaks from the return value. In theory this should be
slightly faster, as a step of the calculation is skipped, and since the
peak digests no longer need to be read from the database.