perf: faster `count_leaves` and `prove_membership_async` for MMRA
With this commit the DB is only accessed once when counting leaves, whereas
earlier, a helper function that read multiple digests from the MMRA was used.

Also changes index arithmetic for getting the relevant node indices for an
MMR authentication path slightly, and then reads all the digests in one
go instead of in a loop.

Good speedup seen from this commit:
- batch-mutation benchmark is 10 % faster, regular mutation is 20 % faster.