perf: make Block::hash() 0(1) via precompute
Previously the block digest was being recomputed each time
`Block::hash()` is called.

This commit changes this so that the hash is computed exactly once for
each modification of Block.  `Block::hash()` is now O(1).

In order to enforce this, the `readonly` crate is used to make Block
fields publically readable, but not writeable. Thus block internals
cannot be modified without using methods on Block.  Every method
that modifies Block now (re-)computes hash and updates Block::digest.

Changes:
 * add dependency on crate `readonly`
 * add a (non-serialized) field `Block::digest`.
 * return `Block::digest` from `Block::hash()`
 * compute digest in `Block::new()`
 * compute digest when deserializing
 * compute digest in `Block::from::<TransferBlock>()`
 * compute digest any time a block is modified.
 * adds tests for each of the above situations.
 * adds `Block::set_header_nonce()` now used by mining loop
 * adapts existing code (mainly tests) to the changes in Block