archival-state: Always set stored block as tip
Always set the new block we are storing to disk and indexing in the
database as the tip. Previously, we would check if this block exceeded
the proof of work family number of the current tip but this
responsibility is now delegated to the caller.

This is needed to fix a bug in `store_block_internal_worker` where the
value `previous_ms_accumulator` is set incorrectly whenever there is a
fork, i.e. whenever `store_block_internal_worker` receives a block that
is not the child of what the block currently considers the tip. In order
to set `previous_ms_accumulator`, we need to know that the block we are
storing now is considered the tip (by the `ArchivalState`) *and* we need
a method to fetch the parent of the tip after this function
`write_block_as_tip` has been called.

Cf. #143.