Always consider newly added block as tip
When the global state is updated with a new block, that new block is
always considered the new tip. This change is reflected in a new name
for this function also. This allows us to correctly handle state updates
in the case of forks. Since the block is no longer carrying around the
previous state of the mutator set accumulator, we have to get that state
from somewhere else. We elect to get it from the archival state which
can identify the parent of the current tip.

Cf. #143