mempool: Clear all transactions on reorganizations
Handling reorganizations in the mempool is complicated. Not only does
the mutator set data need to be rolled back, but that is also the case
for the Triton VM proofs. The easiest solution would probably be to keep
track of a few proofs for each transaction in the mempool such that roll
backs that are only a few blocks deep could be handled.

Care also has to be given to repopulating the mempool with transactions
that were included in the blocks that are being rolled back. These might
have to go into the mempool again.

The current solution, though, is simply to clear the mempool in the case
of a reorganization. This commit includes a test that this is handled
in a sane way, and that the mempool does not get polluted with incorrect/
deprecated/unsynces witness data.

This closes #145.